pyproject.toml
[tool.poetry]
name = "dntt"
version = "0.1.0"
description = "Dump Network Topology Tool"
authors = ["Jiajie Chen <c@jia.je>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
